Move all graphie resource loading/caching into graphie-utils.ts
